Refactor: main 파일에 구현되어 있던 ContactManager 클래스를 새로운 파일로 분리하여 main에 실행부만 남김
